@@ -18,17 +18,17 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  google_fonts: ^2.0.0
-  flutter_screenutil: ^5.0.0-nullsafety.11
-  http: ^0.13.1
+  google_fonts: ^2.1.0
+  flutter_screenutil: ^5.0.0+2
+  http: ^0.13.3
   supercharged: ^2.0.0
-  simple_animations: ^3.0.3
-  pedantic: ^1.11.0
+  simple_animations: ^3.1.1
+  pedantic: ^1.11.1
   intl: ^0.17.0
   flutter_neumorphic: ^3.1.0
 
