@@ -5,15 +5,15 @@ class ThemeSwitcher extends InheritedWidget {
       data; // We'll use ThemeSwitcher to get access to the current state of ThemeSwitcherWidget
 
   const ThemeSwitcher({
-    Key key,
-    @required this.data,
-    @required Widget child,
+    Key? key,
+    required this.data,
+    required Widget child,
   })  : assert(child != null),
         super(key: key, child: child);
 
   static _ThemeSwitcherWidgetState of(BuildContext context) {
     //This method returns the current state of the ThemeSwitcherWidget. This will be used down the tree
-    return (context.dependOnInheritedWidgetOfExactType<ThemeSwitcher>())
+    return context.dependOnInheritedWidgetOfExactType<ThemeSwitcher>()!
         .data;
   }
 
@@ -28,7 +28,7 @@ class ThemeSwitcherWidget extends StatefulWidget {
   final Widget
       child; // child to which this boolean variable should be propagated upon change. This will be our app in this case
 
-  ThemeSwitcherWidget({Key key, this.initialDarkModeOn, this.child})
+  ThemeSwitcherWidget({Key? key, required this.initialDarkModeOn, required this.child})
       : assert(initialDarkModeOn != null),
         assert(child != null),
         super(key: key);
