@@ -144,19 +144,19 @@ class _InfoScreenState extends State<InfoScreen> {
                 crossAxisAlignment: CrossAxisAlignment.stretch,
                 children: [
                   Text(
-                    experienceModel.companyName,
+                    experienceModel.companyName!,
                     textScaleFactor: 2,
                   ),
                   SizedBox(height: 5,),
                   Text(
-                    experienceModel.description,
+                    experienceModel.description!,
                     textScaleFactor: 1.5,
                     style: Theme.of(context).textTheme.caption,
                   ),
                   SizedBox(height: 5,),
                   Text(
-                    experienceModel.marks!=null ? '${DateFormat('yyyy/MM/dd').format(experienceModel.endDate)}' :
-                    '${DateFormat('yyyy/MM/dd').format(experienceModel.startDate)} - ${experienceModel.endDate != null ? DateFormat('yyyy/MM/dd').format(experienceModel.endDate) : 'Present'}',
+                    experienceModel.marks!=null ? '${DateFormat('yyyy/MM/dd').format(experienceModel.endDate!)}' :
+                    '${DateFormat('yyyy/MM/dd').format(experienceModel.startDate!)} - ${experienceModel.endDate != null ? DateFormat('yyyy/MM/dd').format(experienceModel.endDate!) : 'Present'}',
                     style: Theme.of(context).textTheme.caption,
                   ),
                 ],
@@ -198,8 +198,8 @@ class _InfoScreenState extends State<InfoScreen> {
               child: Center(
                 child: Text(
                   experienceModel.marks!=null
-                      ? experienceModel.marks
-                      : '${_getDiff(experienceModel.startDate, experienceModel.endDate ?? DateTime.now())}',
+                      ? experienceModel.marks!
+                      : '${_getDiff(experienceModel.startDate!, experienceModel.endDate ?? DateTime.now())}',
                   style: TextStyle(
                     fontSize: 12,
                     color: Colors.white,
